feat : 수익률 출력 기능 구현

+++ src/main/java/lotto/domain/Won.java
@@ -3,9 +3,9 @@ package lotto.domain;
 import java.util.Objects;
 
 public class Won {
-    private final int amount;
+    private final long amount;
 
-    public Won(int amount) {
+    public Won(long amount) {
         this.amount = amount;
     }
